+++ frontend/app/ui/form/SelectVocabularyItemForm/SectionItems/index.tsx
@@ -1,0 +1,11 @@
+import Item from "../Item";
+import { SectionProps } from "../Section";
+import styles from '../styles.module.css';
+
+export default async function SectionItems({ items }: Pick<SectionProps, 'items'>) {
+  return (
+    <div className={styles.sectionItemsContainer}>
+      {items.map(item => <Item key={item.id.toString()} item={item} />)}
+    </div>
+  );
+}
